+++ raiden-dapp/CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - [#218] Matrix availability check to address inputs.
 - [#585] Pending transfer badge icon
+- [#930] Display complete address on hover
 
 ### Changed
 - User can now dismiss/hide the transfer progress dialog.
@@ -24,6 +25,7 @@
 [#585]: https://github.com/raiden-network/light-client/issues/585
 [#642]: https://github.com/raiden-network/light-client/issues/642
 [#715]: https://github.com/raiden-network/light-client/issues/715
+[#930]: https://github.com/raiden-network/light-client/issues/930
 
 ## [0.2] - 2019-11-29
 ### Added